+++ internautenb2binfo/translations/fr.php
@@ -1,0 +1,23 @@
+<?php
+// French translations for the Internauten B2B Info module.
+// Copyright (c) 2026 die.internauten.ch GmbH
+// License: MIT
+
+global $_MODULE;
+$_MODULE = array();
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_997e7cd3c8f4d8a86a9d755c24f3fdfe'] = 'Internauten B2B Info';
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_ce87ff5397060ac26b59e43540b505f7'] = 'Afficher un texte personnalisé sur les pages produit pour certains groupes de clients';
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_bb8956c67b82c7444a80c6b2433dd8b4'] = 'Voulez-vous vraiment désinstaller ce module ?';
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_19636395066a245172dc03e21c861ca4'] = 'ID de groupe invalide';
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_462390017ab0938911d2d4e964c0cab7'] = 'Paramètres enregistrés avec succès';
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_f4f70727dc34561dfde1a3c529b6205c'] = 'Paramètres';
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_7a01a554b2fe1ebe7f033013913d66b3'] = 'Module actif';
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_8110eae95ded27f8d4c7dd6ea6e4f5d1'] = 'Activer ou désactiver le module';
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_00d23a76e43b46dae9ec7aa9dcbebb32'] = 'Actif';
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_b9f5c797ebbf55adccdd8539a65a0241'] = 'Inactif';
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_08b02383dce834815c99539f02f0fd89'] = 'Groupe revendeurs';
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_9e43df7c6ddc2519bda266ddc1a812fb'] = 'Sélectionnez le groupe attribué aux revendeurs.';
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_c9cc8cce247e49bae79f15173ce97354'] = 'Enregistrer';
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_89850fde9cbf081b89b5cfcc5b940ff0'] = 'Aucun prix spécial disponible pour les revendeurs ! Demandez une offre.';
+$_MODULE['<{internautenb2binfo}prestashop>displayproductpriceblock_3aa0f3cc40b36a33a804075241ab3c56'] = 'Cet article n\'a pas de prix pour les revendeurs. Ajoutez l\'article dans la quantité souhaitée au panier et choisissez « Demander une offre » comme moyen de paiement lors de la commande.';
+$_MODULE['<{internautenb2binfo}prestashop>displayproductpriceblock_e545bb74dcb36c1382aee29f7b829659'] = 'Prix d\'origine :';