+++ internautenb2binfo/translations/it.php
@@ -1,0 +1,23 @@
+<?php
+// Italian translations for the Internauten B2B Info module.
+// Copyright (c) 2026 die.internauten.ch GmbH
+// License: MIT
+
+global $_MODULE;
+$_MODULE = array();
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_997e7cd3c8f4d8a86a9d755c24f3fdfe'] = 'Internauten B2B Info';
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_ce87ff5397060ac26b59e43540b505f7'] = 'Mostra un testo personalizzato nelle pagine prodotto per determinati gruppi di clienti';
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_bb8956c67b82c7444a80c6b2433dd8b4'] = 'Vuoi davvero disinstallare questo modulo?';
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_19636395066a245172dc03e21c861ca4'] = 'ID gruppo non valido';
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_462390017ab0938911d2d4e964c0cab7'] = 'Impostazioni salvate con successo';
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_f4f70727dc34561dfde1a3c529b6205c'] = 'Impostazioni';
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_7a01a554b2fe1ebe7f033013913d66b3'] = 'Modulo attivo';
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_8110eae95ded27f8d4c7dd6ea6e4f5d1'] = 'Attiva o disattiva il modulo';
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_00d23a76e43b46dae9ec7aa9dcbebb32'] = 'Attivo';
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_b9f5c797ebbf55adccdd8539a65a0241'] = 'Inattivo';
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_08b02383dce834815c99539f02f0fd89'] = 'Gruppo rivenditori';
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_9e43df7c6ddc2519bda266ddc1a812fb'] = 'Seleziona il gruppo assegnato ai rivenditori.';
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_c9cc8cce247e49bae79f15173ce97354'] = 'Salva';
+$_MODULE['<{internautenb2binfo}prestashop>internautenb2binfo_89850fde9cbf081b89b5cfcc5b940ff0'] = 'Nessun prezzo speciale disponibile per i rivenditori! Richiedi un\'offerta.';
+$_MODULE['<{internautenb2binfo}prestashop>displayproductpriceblock_3aa0f3cc40b36a33a804075241ab3c56'] = 'Questo articolo non ha un prezzo per i rivenditori. Aggiungi l\'articolo al carrello nella quantità desiderata e, al momento dell\'ordine, scegli «Richiedi un\'offerta» come metodo di pagamento.';
+$_MODULE['<{internautenb2binfo}prestashop>displayproductpriceblock_e545bb74dcb36c1382aee29f7b829659'] = 'Prezzo originale:';